refactor(mba): keep task semantics out of shared terms
