+++ src/unilab/conf/ppo/task/microduck_ground_pick_flat/mjwarp.yaml
@@ -27,13 +27,13 @@ env:
   render_spacing: 2.0
   commands:
     twist:
-      _target_: unilab.envs.mdp.GroundPickPhaseCommandCfg
+      _target_: unilab.tasks.locomotion.microduck.manager_terms.GroundPickPhaseCommandCfg
       period: 4.0
       randomize_phase: true
   metrics:
     root_height:
       _target_: unilab.managers.MetricsTermCfg
-      func: unilab.tasks.locomotion.microduck.manager_terms.root_height_metric
+      func: unilab.envs.mdp.root_height
       reduce: mean
       params:
         asset_cfg:
@@ -42,11 +42,11 @@ env:
   recorders:
     lifecycle:
       _target_: unilab.managers.RecorderTermCfg
-      func: unilab.tasks.locomotion.microduck.manager_terms.MicroduckTraceRecorder
+      func: unilab.envs.mdp.LifecycleCounterRecorder
 
 reward:
-  posture_height:
-    func: unilab.tasks.locomotion.microduck.manager_terms.posture_height_tracking
+  phase_height:
+    func: unilab.tasks.locomotion.microduck.manager_terms.phase_height_tracking
     weight: 0.5
     params:
       command_name: twist
