+++ src/unilab/conf/ppo/task/microduck_sitstand_flat/mjwarp.yaml
@@ -26,7 +26,7 @@ env:
   render_spacing: 2.0
   commands:
     twist:
-      _target_: unilab.envs.mdp.SitStandCommandCfg
+      _target_: unilab.tasks.locomotion.microduck.manager_terms.SitStandCommandCfg
       sit_prob: 0.5
       ramp_s: 2.0
       sit_z: 0.060
@@ -34,7 +34,7 @@ env:
   metrics:
     root_height:
       _target_: unilab.managers.MetricsTermCfg
-      func: unilab.tasks.locomotion.microduck.manager_terms.root_height_metric
+      func: unilab.envs.mdp.root_height
       reduce: mean
       params:
         asset_cfg:
@@ -43,7 +43,7 @@ env:
   recorders:
     lifecycle:
       _target_: unilab.managers.RecorderTermCfg
-      func: unilab.tasks.locomotion.microduck.manager_terms.MicroduckTraceRecorder
+      func: unilab.envs.mdp.LifecycleCounterRecorder
 
 reward:
   posture_height:
