+++ src/unilab/conf/sac/task/microduck_velocity_flat/base.yaml
@@ -189,7 +189,7 @@ env:
         lin_vel_y: [-0.3, 0.3]
         ang_vel_z: [-1.0, 1.0]
     head_pose:
-      _target_: unilab.tasks.locomotion.microduck.manager_terms.UniformVectorCommandCfg
+      _target_: unilab.envs.mdp.UniformPoseCommandCfg
       resampling_time_range: [2.0, 5.0]
       ranges:
         - [-0.05, 0.05]
@@ -197,7 +197,7 @@ env:
         - [-0.07, 0.07]
         - [-0.015, 0.015]
     body_pose:
-      _target_: unilab.tasks.locomotion.microduck.manager_terms.UniformVectorCommandCfg
+      _target_: unilab.envs.mdp.UniformPoseCommandCfg
       resampling_time_range: [2.0, 5.0]
       ranges:
         - [-0.005, 0.005]
