+++ src/unilab/envs/mdp/commands/__init__.py
@@ -2,14 +2,6 @@
 
 from unilab.envs.mdp.commands.pose_command import UniformPoseCommand as UniformPoseCommand
 from unilab.envs.mdp.commands.pose_command import UniformPoseCommandCfg as UniformPoseCommandCfg
-from unilab.envs.mdp.commands.posture_command import (
-    GroundPickPhaseCommand as GroundPickPhaseCommand,
-)
-from unilab.envs.mdp.commands.posture_command import (
-    GroundPickPhaseCommandCfg as GroundPickPhaseCommandCfg,
-)
-from unilab.envs.mdp.commands.posture_command import SitStandCommand as SitStandCommand
-from unilab.envs.mdp.commands.posture_command import SitStandCommandCfg as SitStandCommandCfg
 from unilab.envs.mdp.commands.velocity_command import (
     UniformVelocityCommand as UniformVelocityCommand,
 )
@@ -18,10 +10,6 @@ from unilab.envs.mdp.commands.velocity_command import (
 )
 
 __all__ = [
-    "GroundPickPhaseCommand",
-    "GroundPickPhaseCommandCfg",
-    "SitStandCommand",
-    "SitStandCommandCfg",
     "UniformPoseCommand",
     "UniformPoseCommandCfg",
     "UniformVelocityCommand",
